+++ domain/src/main/java/ru/dpopkov/knownics/domain/KeyTermRepository.java
@@ -1,0 +1,4 @@
+package ru.dpopkov.knownics.domain;
+
+public interface KeyTermRepository extends BaseRepository<KeyTerm> {
+}