+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/SourceRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.answer;
+
+import ru.dpopkov.knownics.domain.BaseRepository;
+
+public interface SourceRepository extends BaseRepository<Source> {
+}