+++ domain/src/main/java/ru/dpopkov/knownics/domain/deck/DeckRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.deck;
+
+import ru.dpopkov.knownics.domain.BaseRepository;
+
+public interface DeckRepository extends BaseRepository<Deck> {
+}