+++ domain/src/main/java/ru/dpopkov/knownics/domain/deck/FlashCardRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.deck;
+
+import ru.dpopkov.knownics.domain.BaseRepository;
+
+public interface FlashCardRepository extends BaseRepository<FlashCard> {
+}