+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/QCollectionRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.question;
+
+import ru.dpopkov.knownics.domain.BaseRepository;
+
+public interface QCollectionRepository extends BaseRepository<QCollection> {
+}