+++ domain/src/main/java/ru/dpopkov/knownics/domain/user/AppUserRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.user;
+
+import ru.dpopkov.knownics.domain.BaseRepository;
+
+public interface AppUserRepository extends BaseRepository<AppUser> {
+}